+++ packages/cms/lib/modules/resource-form-widgets/lib/fields.js
@@ -1,4 +1,11 @@
 const fields = [
+    {
+      name: 'formName',
+      type: 'string',
+      label: 'Unique form name',
+      help: 'This form id must be unique on this site and must be lowercase with no whitespaces',
+      required: true
+    },
      {
         name: 'redirect',
         type: 'string',
@@ -26,6 +33,38 @@ const fields = [
               showFields: ['dynamicFormSections']
             },
         ]
+    },
+    {
+      name: 'formVisibility',
+      type: 'select',
+      label: 'Who can view this form?',
+      choices: [
+        {
+          value: 'user',
+          label: "Only users (and admin users)",
+        },
+        {
+          value: 'always',
+          label: "Everyone"
+        },
+      ],
+      def: 'user'
+    },
+    {
+      name: 'shouldDisplayUserName',
+      type: 'boolean',
+      label: 'Show user name in form header?',
+      choices: [
+        {
+          label: 'Yes',
+          value: true,
+        },
+        {
+          label: 'No',
+          value: false,
+        }
+      ],
+      def: true
     },
     {
       name: 'hideAdminAfterPublicAction',
@@ -77,6 +116,16 @@ const fields = [
                       value: 'radio',
                       label: "Radio",
                       showFields: ['fieldKey', 'fieldRequired', 'fieldOptions', 'notExtraDataKey']
+                  },
+                  {
+                      value: 'checkbox',
+                      label: 'Checkbox',
+                      showFields: ['fieldKey', 'fieldRequired', 'fieldOptions', 'notExtraDataKey']
+                  },
+                  {
+                    value: 'select',
+                    label: 'Select',
+                    showFields: ['fieldKey', 'fieldRequired', 'fieldOptions', 'notExtraDataKey']
                   },
                   {
                       value: 'text',
@@ -787,6 +836,102 @@ const fields = [
        type: 'string',
        label: 'Text for button to save',
     },
+    {
+      name: 'confirmationEnabledUser',
+      type: 'boolean',
+      label: 'Confirmation to user',
+      help: 'Only implemented for resource `submissions` at the moment, other resource will use the "default" notification flow',
+      choices: [
+        {
+          value: 1,
+          label: "Yes",
+          showFields: ['confirmationTemplateNameUser', 'confirmationSubjectUser', 'confirmationEmailFieldUser', 'confirmationTemplateNameUser', 'confirmationEmailContentUser']
+        },
+        {
+          value: 0,
+          label: "No"
+        }
+      ]
+    },
+    {
+      name: 'confirmationTemplateNameUser',
+      type: 'select',
+      label: 'User template name',
+      choices: [
+        {
+          label: 'Submission default',
+          value: 'submission_default.njk'
+        },
+        {
+          label: 'Blanco template',
+          value: 'blanco.njk'
+        }
+      ]
+    },
+    {
+      name: 'confirmationSubjectUser',
+      type: 'string',
+      label: 'Confirmation subject to user',
+    },
+    {
+      name: 'confirmationEmailFieldUser',
+      type: 'string',
+      label: 'Confirmation email field',
+    },
+    {
+      name: 'confirmationEmailContentUser',
+      type: 'string',
+      label: 'Confirmation email content',
+      textarea: true
+    },
+    {
+      name: 'confirmationEnabledAdmin',
+      type: 'boolean',
+      label: 'Confirmation to admin',
+      help: 'Only implemented for resource `submissions` at the moment, other resource will use the "default" notification flow',
+      choices: [
+        {
+          value: 1,
+          label: "Yes",
+          showFields: ['confirmationTemplateNameAdmin', 'confirmationSubjectAdmin', 'confirmationEmailFieldAdmin', 'confirmationTemplateNameAdmin', 'confirmationEmailContentAdmin']
+        },
+        {
+          value: 0,
+          label: "No"
+        }
+      ]
+    },
+    {
+      name: 'confirmationTemplateNameAdmin',
+      type: 'select',
+      label: 'Admin template name',
+      choices: [
+        {
+          label: 'submission default',
+          value: 'submission_default.njk'
+        },
+        {
+          label: 'Blanco template',
+          value: 'blanco.njk'
+        }
+      ]
+    },
+    {
+      name: 'confirmationSubjectAdmin',
+      type: 'string',
+      label: 'Confirmation subject to Admin',
+    },
+    {
+      name: 'confirmationEmailFieldAdmin',
+      type: 'string',
+      label: 'Admin email field',
+    },
+    {
+      name: 'confirmationEmailContentAdmin',
+      type: 'string',
+      label: 'Admin email content',
+      textarea: true
+    }
 ];
 
 module.exports = fields;